+++ scripts/ci-unit-tests.sh
@@ -1,6 +1,7 @@
 command_file="build/tmp/unit-tests-command.sh"
 
-./gradlew test-debug && \
-chmod +x $command_file && \
-./$command_file && \
-rm $command_file
+#./gradlew test-debug && \
+#chmod +x $command_file && \
+#./$command_file && \
+#rm $command_file
+./gradlew testCoverage